CareChainX/backend/pyproject.toml
[project]
name = "carechain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "celery>=5.5.3",
    "dj-database-url>=3.0.1",
    "django>=5.2.5",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "faker>=37.5.3",
    "geopy>=2.4.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "python-telegram-bot[job-queue]>=22.3",
    "razorpay>=1.4.2",
    "redis>=6.4.0",
]
